+++ src/lib.rs
@@ -6,8 +6,8 @@
 )]
 //! A Rust api client for [Trovo](https://trovo.live).
 //!
-//! Currently supports connecting to chat and looking up users via username, more will be added as
-//! the crate develops.
+//! Currently supports connecting to chat, sending messages, and looking up users via username, more
+//! will be added as the crate develops.
 //!
 //! ## Example
 //!